Update CSS to fix lack of styling with no category
CSS did not style output if the Prayer Item had no category.  So add additional CSS for when there is no category on the outer div of each prayer item.  Leave in the original style selectors so that a user can separate them at a later stage to distinguish between no category or categorised prayer items if they wish.